backend/requirements.txt
# FastAPI 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# 文件处理
python-docx==1.1.0
PyMuPDF==1.23.8
Pillow==10.2.0
pdf2image==1.17.0
pywin32==306  # Windows: 用于处理 .doc 文件

# HTTP 客户端
httpx==0.26.0

# GPU 显存（侧栏 /health/services；NVML，不依赖 PATH 中的 nvidia-smi）
nvidia-ml-py>=12.535.133
aiofiles==23.2.1

# 智谱AI SDK
zhipuai>=2.0.0

# HaS Image（YOLO11 微服务 has_image_server.py，端口 8081）
ultralytics>=8.3.0

# OCR（PaddleOCR-VL 微服务 ocr_server.py）
# 推荐：先安装 GPU 版 Paddle，再装本文件，避免默认装上 CPU 版 paddlepaddle。
# Windows + NVIDIA：在项目根目录执行 scripts\install_paddle_gpu.ps1，然后:
#   pip install -r requirements.txt
# 验证: python -c "import paddle; print(paddle.get_device())"  # 应为 gpu:0
paddleocr[doc-parser]>=2.7.0

# 工具库
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 开发工具
python-dotenv==1.0.0
pytest>=8.0.0
